+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Add affected and fixed version information in report.
+- Introduce `fixedAt` DSL to define a vulnerability report as fixed.
 
 ## [0.6.0] - 2025-03-15